Refactor function definitions
