+++ action.yml
@@ -28,8 +28,6 @@ inputs:
 outputs:
   response: # output will be available to future steps
     description: 'The response'
-  test:
-    description: 'output'
 runs:
   using: 'node16'
   main: 'dist/index.js'
