+++ package.json
@@ -26,6 +26,7 @@
   "homepage": "https://github.com/actions/javascript-action#readme",
   "dependencies": {
     "@actions/core": "^1.10.0",
+    "@actions/github": "^5.1.1",
     "axios": "^1.4.0",
     "dotenv": "^16.0.3",
     "openai": "^3.2.1"
