+++ src/prompt.js
@@ -1,13 +1,14 @@
 const fetchGithubActivities = require('./github-api')
 
 module.exports = async function generatePrompt(
+      username,
       persona,
       githubInput,
       action,
       tone,
       outputLength) {
         
-       let githubUsage = await fetchGithubActivities("username")
+       let githubUsage = await fetchGithubActivities(username)
        let prompt =  `
         Assume you are ${persona}. I am a github user who has
         ${githubUsage}. Your response should not be negative and discouraging.
